power: add FAN53555 family support
This adds a driver for the FAN53555 family of regulators and wraps it
in a PMIC implementation.

While these devices support a 'normal' and 'suspend' mode (controlled
via an external pin) to switch between two programmable voltages, this
incarnation of the driver assumes that the device is always operating
in 'normal' mode.

Only setting/reading the programmed voltage is supported at this time
and the following device functionality remains unsupported:
  - switching the selected voltage (via a GPIO)
  - disabling the voltage output via software-control
This matches the functionality of the Linux driver.

Tested on a RK3399-Q7 (with 'option 5' devices): setting voltages from
the U-Boot shell and verifying output voltages on the board.

Signed-off-by: Philipp Tomsich <philipp.tomsich@theobroma-systems.com>
Tested-by: Klaus Goger <klaus.goger@theobroma-systems.com>

Series-version: 5

Series-changes: 5
- introduced a PMIC (as requested by SJG) to wrap the single regulator

Series-changes: 4
- fix issues introduced when updating for review comments (sorry for
  this: my submit-branch had diverged from the WIP-branch I used for
  testing and I didn't notice)

Series-changes: 3
- update for review comments

Series-changes: 2
- adapted documentation on the device-tree binding from Linux