bootcount: add uclass for bootcount
The original bootcount methods do not provide an interface to DM and
rely on a static configuration for I2C devices (e.g. bus, chip-addr,
etc. are configured through defines statically).  On a modern system
that exposes multiple devices in a DTS-configurable way, this is less
than optimal and a interface to DM-based devices will be desirable.

This adds a simple driver that is DM-aware and configurable via DTS.
If ambiguous (i.e. multiple bootcount-devices are present) the
/chosen/u-boot,bootcount-device property can be used to select one
bootcount device.

Initially, this provides support for the following DM devices:
 * RTC devices

Signed-off-by: Philipp Tomsich <philipp.tomsich@theobroma-systems.com>
Tested-by: Klaus Goger <klaus.goger@theobroma-systems.com>
Series-cc: Klaus Goger <klaus.goger@theobroma-systems.com>

Series-version: 2
Series-changes: 2
- changed to provide a UCLASS-based implementation, as requested by
  SJG in his earlier review
- split off the RV3029 driver into a separate series

Cover-letter:
Update RV3029 driver to DM and add DM-backed bootcount support

On one of our application-specific carrier boards, a MicroCrystal
RV3029 acts as an off-module battery-backed RTC for the RK3399-Q7.
The RV3029 is intended both to provide RTC services (to Linux) and
to store the bootcount in its battery-backed SRAM section.

To support this use-case, this series adds the following:
 * replaces the existing RV3029 driver by a DM-based one (note that the
   existing driver appears unused and didn't even have a Kconfig entry)
   that closely mirrors its incarnation in Linux 4.17
 * adds a bootcount-method interfacing back into DM devices (implementing
   support for the RTC uclass as of now).

END